backend/requirements.txt
fastapi
pydantic
groq
uvicorn
pandas
python-dotenv